backend/requirements.txt
Flask
Flask-Cors
obsidian-apo
pandas
numpy